+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.11.7
+
+* `Common`:
+  * New abstractions `SimpleMapper` and `SimpleSuspendableMapper`
+* `Repos`:
+  * `Common`:
+    * Add mappers for `CRUDRepo`
+
 ## 0.11.6
 
 * `FSM`: